[bphh-635] Refactor requirement-field-display sub components to individual files and organize them under folder
