+++ app/frontend/components/domains/requirements-library/requirement-field-display/contact-field-item-display.tsx
@@ -1,0 +1,31 @@
+import React from "react"
+import { getRequirementContactFieldItemLabel } from "../../../../constants"
+import { ERequirementContactFieldItemType, ERequirementType } from "../../../../types/enums"
+import { RequirementFieldDisplay, TRequirementFieldDisplayProps } from "./index"
+
+interface IContactFieldItemDisplayProps {
+  contactFieldItemType: ERequirementContactFieldItemType
+}
+
+export function ContactFieldItemDisplay({ contactFieldItemType }: IContactFieldItemDisplayProps) {
+  const defaultProps: Partial<TRequirementFieldDisplayProps> = {
+    label: getRequirementContactFieldItemLabel(contactFieldItemType),
+    requirementType: ERequirementType.text,
+  }
+
+  const propsByType = {
+    [ERequirementContactFieldItemType.email]: {
+      requirementType: ERequirementType.email,
+    },
+    [ERequirementContactFieldItemType.phone]: {
+      requirementType: ERequirementType.phone,
+    },
+    [ERequirementContactFieldItemType.address]: {
+      requirementType: ERequirementType.address,
+    },
+    [ERequirementContactFieldItemType.organization]: {
+      labelProps: { maxW: "full" },
+    },
+  }
+  return <RequirementFieldDisplay {...defaultProps} {...propsByType[contactFieldItemType]} />
+}
